:truck: Move CLI app to package app
