+++ modules/cli/src/main/kotlin/datamaintain/cli/app/ListExecutedScripts.kt
@@ -1,6 +1,6 @@
-package datamaintain.cli
+package datamaintain.cli.app
 
-import datamaintain.cli.utils.logger
+import datamaintain.cli.app.utils.logger
 import datamaintain.core.Datamaintain
 import datamaintain.core.config.DatamaintainConfig
 import java.util.*
