+++ modules/test/src/main/kotlin/datamaintain/test/AppSubProcessUtils.kt
@@ -18,7 +18,7 @@ fun execAppInSubprocess(args: List<String> = emptyList(), jvmArgs: List<String> 
     val javaHome = System.getProperty("java.home")
     val javaBin = javaHome + File.separator + "bin" + File.separator + "java"
     val classpath = System.getProperty("java.class.path")
-    val className = "datamaintain.cli.AppKt"
+    val className = "datamaintain.cli.app.AppKt"
 
     val command = ArrayList<String>()
     command.add(javaBin)
